requirements.txt
discord.py==2.2.2
openai==0.27.2
python-dotenv==1.0.0
youtube_dl==2021.12.17
